prefer structs over tuples for day 2

+++ include/Day02.hpp
@@ -26,7 +26,7 @@ namespace day02 {
         switch (c) {
           case 'u': return dir::up;
           case 'd': return dir::down;
-          default: return dir::forward;
+          default:  return dir::forward;
         }
       } (dir[0]);
       ctx.range() = std::move(r.range());
@@ -37,5 +37,5 @@ namespace day02 {
   constexpr const int id = 2;
   using parsed = std::vector<data>;
   using result1 = int;
-  using result2 = long;
+  using result2 = int;
 }
